requirements.txt
# LangChain ecosystem - using compatible versions
langchain
langchain-openai
langchain-community
langgraph

# Database
asyncpg==0.29.0

# Core dependencies
pydantic==2.7.1
python-dotenv==1.0.1

# API framework
fastapi==0.111.0
uvicorn==0.29.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

# Optional: for better logging and monitoring
structlog==24.1.0